moving to PEP-8 style, take files from the __init__ and put them in more logical, individual, files
