+++ src/DicomRTTool/Viewer.py
@@ -1,0 +1,62 @@
+__author__ = 'Brian M Anderson'
+# Created on 12/31/2020
+import numpy as np
+import copy
+import matplotlib.pyplot as plt
+
+
+def plot_scroll_Image(x):
+    '''
+    :param x: input to view of form [rows, columns, # images]
+    :return:
+    '''
+    if x.dtype not in ['float32','float64']:
+        x = copy.deepcopy(x).astype('float32')
+    if len(x.shape) > 3:
+        x = np.squeeze(x)
+    if len(x.shape) == 3:
+        if x.shape[0] != x.shape[1]:
+            x = np.transpose(x,[1,2,0])
+        elif x.shape[0] == x.shape[2]:
+            x = np.transpose(x, [1, 2, 0])
+    fig, ax = plt.subplots(1, 1)
+    if len(x.shape) == 2:
+        x = np.expand_dims(x,axis=-1)
+    tracker = IndexTracker(ax, x)
+    fig.canvas.mpl_connect('scroll_event', tracker.onscroll)
+    return fig,tracker
+    #Image is input in the form of [#images,512,512,#channels]
+
+
+class IndexTracker(object):
+    def __init__(self, ax, X):
+        self.ax = ax
+        ax.set_title('use scroll wheel to navigate images')
+
+        self.X = X
+        rows, cols, self.slices = X.shape
+        self.ind = np.where((np.min(self.X,axis=(0,1))!= np.max(self.X,axis=(0,1))))[-1]
+        if len(self.ind) > 0:
+            self.ind = self.ind[len(self.ind)//2]
+        else:
+            self.ind = self.slices//2
+
+        self.im = ax.imshow(self.X[:, :, self.ind],cmap='gray')
+        self.update()
+
+    def onscroll(self, event):
+        print("%s %s" % (event.button, event.step))
+        if event.button == 'up':
+            self.ind = (self.ind + 1) % self.slices
+        else:
+            self.ind = (self.ind - 1) % self.slices
+        self.update()
+
+    def update(self):
+        self.im.set_data(self.X[:, :, self.ind])
+        self.ax.set_ylabel('slice %s' % self.ind)
+        self.im.axes.figure.canvas.draw()
+
+
+if __name__ == '__main__':
+    pass
